refactor(api): Changed implementation of DbContextFake to behave more like in a real world scenario (#54)
WIP refactor(api): Change entity and controller names

+++ src/Intranet.API/Intranet.API.UnitTests/Fakes/DbContextFake.cs
@@ -6,84 +6,44 @@ using System.Linq;
 using System.Text;
 using System.Reflection;
 using System.Runtime.CompilerServices;
+using Microsoft.EntityFrameworkCore.ChangeTracking;
 
 namespace Intranet.API.UnitTests.Fakes
 {
   public static class DbContextFake
   {
-    public static TContext GetDbContext<TContext>([CallerMemberName] string callerName = "")
-      where TContext : DbContext
-    {
-      var data = SingleItem(String.Empty, new List<object>());
-      return GetDbContext<TContext>(data, callerName);
-    }
-
-    public static TContext GetDbContext<TContext>(string property, object obj, [CallerMemberName] string callerName = "")
-      where TContext : DbContext
-    {
-      var listType = typeof(List<>).MakeGenericType(obj.GetType());
-      var listOfEntities = (IList)Activator.CreateInstance(listType);
-      listOfEntities.Add(obj);
-
-      return GetDbContext<TContext>(new List<(string, IEnumerable<object>)> { (property, listOfEntities as IEnumerable<object>) }, callerName);
-    }
-
-    public static TContext GetDbContext<TContext>(string property, IEnumerable<object> obj, [CallerMemberName] string callerName = "")
-      where TContext : DbContext
-    {
-      var data = SingleItem(property, obj);
-      return GetDbContext<TContext>(data, callerName);
-    }
-
-    public static TContext GetDbContext<TContext>(IEnumerable<(string property, IEnumerable<object> obj)> data, [CallerMemberName] string callerName = "")
+    /// <summary>
+    /// Creates a new DbContext that is using an In Memory-database
+    /// </summary>
+    /// <typeparam name="TContext"></typeparam>
+    /// <param name="databaseName">If not set this will be the name of the calling method</param>
+    /// <returns></returns>
+    public static TContext GetDbContext<TContext>([CallerMemberName] string databaseName = "")
       where TContext : DbContext
     {
       var options = new DbContextOptionsBuilder<TContext>()
-          .UseInMemoryDatabase(databaseName: callerName)
+          .UseInMemoryDatabase(databaseName)
           .Options;
 
-      var context = (TContext)Activator.CreateInstance(typeof(TContext), options);
-
-      var addRangeMethodName = nameof(List<string>.AddRange);
-      var setMethodName = nameof(context.Set);
-
-      foreach (var item in data)
-      {
-        if (String.IsNullOrWhiteSpace(item.property) || item.obj == null)
-        {
-          continue;
-        }
-
-        var typeOfEntity = typeof(TContext)
-          .GetProperty(item.property)
-          .PropertyType
-          .GetGenericArguments()
-          .SingleOrDefault();
-
-        var dbSet = context
-          .GetType()
-          .GetMethod(setMethodName)
-          .MakeGenericMethod(typeOfEntity)
-          .Invoke(context, null);
-
-        dbSet
-          .GetType()
-          .GetMethod(addRangeMethodName, new Type[] { item.obj.GetType() })
-          .Invoke(dbSet, new[] { item.obj });
-      }
-      context.Database.EnsureDeleted();
-      context.SaveChanges();
-
-      return context;
+      return (TContext)Activator.CreateInstance(typeof(TContext), options);
     }
 
-    #region Private helpers
-
-    private static IEnumerable<(string, IEnumerable<T>)> SingleItem<T>(string property, IEnumerable<T> data)
+    /// <summary>
+    /// Seeds the database
+    /// </summary>
+    /// <typeparam name="TContext"></typeparam>
+    /// <param name="action"></param>
+    /// <param name="ensureDeleted">If true the database will be deleted if it exists</param>
+    /// <param name="databaseName">If not set this will be the name of the calling method</param>
+    public static void SeedDb<TContext>(Action<TContext> action, bool ensureDeleted = false, [CallerMemberName] string databaseName = "")
+      where TContext : DbContext
     {
-      return new List<(string, IEnumerable<T>)> { (property, data) };
+      using (var context = GetDbContext<TContext>(databaseName))
+      {
+        if (ensureDeleted) context.Database.EnsureDeleted();
+        action(context);
+        context.SaveChanges();
+      }
     }
-
-    #endregion
   }
 }
